refactor/EV-780/refactoring files location and names
