+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   hooks_riverpod: ^1.0.4
   decimal: ^2.3.0
   http: ^0.13.5
+  device_preview: ^1.1.0
 
 dev_dependencies:
   flutter_test:
